% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_faa_aids.R
\name{load_aidcodes}
\alias{load_aidcodes}
\title{Load AID code table files}
\usage{
load_aidcodes(code_dir, code_file_pattern)
}
\arguments{
\item{code_dir}{Directory containing AID code table files}

\item{code_file_pattern}{Regex pattern for AID code table filenames}
}
\value{
A named vector of vectors, each usable as ... of forcats::fct_recode()
}
\description{
This function reads all TSV format text files in a specified directory that
match a specified regular expression pattern. It expects the regex pattern
to have a capture group that is unique to each file. It expects each file to
contain a column of codes named CODE and a column of descriptions named
LONG DESCRIPTION. It returns a named vector (named for the regex capture
group contents) of named vectors, each of which is suitable for passing to
forcats::fct_recode() as the ... \if{html}{\out{<dynamic-dots>}} argument.
}
