% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_faa_aids.R
\name{load_aids}
\alias{load_aids}
\title{Load AIDS data files}
\usage{
load_aids(
  data_dir = system.file("extdata", package = "faa.aids"),
  a_file_pattern = "^a.*[0-9]{2}\\\\.txt",
  e_file_pattern = "^e.*[0-9]{2}\\\\.txt"
)
}
\arguments{
\item{data_dir}{Directory with FAA data files (defaults to package data)}

\item{a_file_pattern}{Regex pattern for FAA AIDS A (main data) filenames}

\item{e_file_pattern}{Regex pattern for FAA AIDS E (remarks) filenames}
}
\value{
A tidyr::tibble of FAA AIDS data
}
\description{
This function reads TSV format text files published by the U.S. Federal
Aviation Administration (FAA) containing data from their Accident/Incident
Data System (AIDS), and it decodes factor values using the FAA's published
code tables. See also documentation for load_aidcodes() and load_farparts().
}
