% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_faa_aids.R
\name{load_farparts}
\alias{load_farparts}
\title{Load a FAR parts file}
\usage{
load_farparts(farparts_file)
}
\arguments{
\item{farparts_file}{Path to the TSV file of FAR parts}
}
\value{
A named vector usable as ... of forcats::fct_recode()
}
\description{
This function reads a TSV format text file containing a table of numbered
parts of the U.S. Federal Aviation Regulations (FAR) and their
corresponding descriptions. It expects FAR part numbers in the first column
and descriptions in the second column, without a header row of column names.
If no file path is specified, it will read the file provided by the package.
It returns a named vector suitable for passing to forcats::fct_recode() as
the ... \if{html}{\out{<dynamic-dots>}} argument.
}
